refactor(game-page): drop redundant Native Linux hint; OS chips now open Metadata
The 'Native Linux runtime available (metadata)' hint under the box art
was saying the same thing the green Linux chip up top already says:
Steam offers a native Linux build. Two visual sources for the same
signal read as noise; the mono-italic hint kept pulling the eye away
from the header layout without adding information.

Retires the hint entirely and moves the click affordance onto the OS
chips themselves (Win / macOS / Linux). Each chip:
  * lights up green when Steam advertises the platform (unchanged)
  * is now a <button> instead of a <span> so keyboard focus works
  * carries a click handler that opens the Metadata modal for that app
  * has an updated title/tooltip explaining what a click reveals

Matches the mental model of 'tap the Linux chip to see when the Linux
build was last updated' -- exactly what the modal answers now that
issue #215 seeded real per-OS depot dates for Hollow Knight.

Also drops the .game-native-linux CSS block, the .tux-icon selector,
and the (metadata) suffix span. Adds hover states to the chips so they
look interactive.

Test guard updated: runtimeHistoryModal.test.js no longer expects the
retired native-hint click wiring; instead asserts the OS chip click
handler is present. The _openRuntimeHistoryModal function itself still
exists (aggregates community reports by run_type) but is no longer
reachable from the header UI -- kept for now in case a future 'report
timing' surface wants it back.
