+++ game-stats.html
@@ -13,7 +13,7 @@
 <link rel="stylesheet" href="css/shared/topbar.css?v=c550d77b">
 <link rel="stylesheet" href="css/shared/site.css?v=619dee5a">
 <link rel="stylesheet" href="css/shared/cards.css?v=0817c80a">
-<link rel="stylesheet" href="css/app/game-header.css?v=2dc3da06">
+<link rel="stylesheet" href="css/app/game-header.css?v=8018fb80">
 <link rel="stylesheet" href="css/app/configs.css?v=90d5bed5">
 <link rel="stylesheet" href="css/app/reports.css?v=6df19441">
 <link rel="stylesheet" href="css/app/modals.css?v=5e3db937">
