+++ tests/runtimeHistoryModal.test.js
@@ -1,15 +1,12 @@
 /**
- * The runtime-history modal is opened by clicking the Native Linux hint on
- * the game page. Rather than pulling in jsdom + the full game-page render
- * we assert the source contains the aggregation invariants that drive it:
- *   - groups by run_type (unknown for legacy null rows)
- *   - counts reports per runtime
- *   - tracks first-seen (min timestamp) + last-updated (max)
- *   - orders canonical runtimes first, unknown last
+ * The runtime-history modal still exists (aggregates community reports
+ * by runtime type) but is no longer wired to the retired Native Linux
+ * hint. Kept here as a source-shape regression guard on the aggregation
+ * so a future refactor cannot silently break the report-timing view.
  *
- * Regression guards on the SQL-ish shape stay in a source-string test
- * because the aggregation lives inline in the game-page module (no
- * import surface to unit test directly).
+ * The Native Linux hint under the artwork was removed as redundant with
+ * the OS chip strip up top; the chips now carry the click affordance for
+ * the Metadata modal.
  */
 
 const fs = require('fs');
@@ -19,11 +16,15 @@ const SRC = fs.readFileSync(
   path.join(__dirname, '..', 'js', 'app', 'components', 'game-page.js'), 'utf8');
 
 describe('runtime history modal (game-page)', () => {
-  test('module exposes _openRuntimeHistoryModal wired from the native badge', () => {
-    // Native badge triggers the modal on click AND on Enter/Space (a11y).
-    expect(SRC).toMatch(/_openRuntimeHistoryModal\(appId, combined\)/);
-    expect(SRC).toMatch(/nativeEl\.addEventListener\('click'/);
-    expect(SRC).toMatch(/nativeEl\.addEventListener\('keydown'/);
+  test('module still defines _openRuntimeHistoryModal even though the native hint is gone', () => {
+    expect(SRC).toMatch(/function _openRuntimeHistoryModal\(appId, combined\)/);
+  });
+
+  test('OS chip strip is the click affordance for the Metadata modal (replaces the retired Native hint)', () => {
+    // Each OS chip now opens the metadata modal; the old .game-native-linux
+    // wiring was redundant with the availability chips.
+    expect(SRC).not.toMatch(/nativeEl\.addEventListener\('click'/);
+    expect(SRC).toMatch(/chip\.addEventListener\('click', \(\) => _openMetadataModal\(appId\)\)/);
   });
 
   test('aggregation buckets on r.runType and falls back to "unknown"', () => {
